32th lesson finished

+++ C kursu/32.Ders-  PointerIdıoms-TypeDefBildirimleri/nutility.h
@@ -3,7 +3,7 @@
 
 
 
-
+#include <stddef.h> // bir kaç macro ve size_t bildirimi için kullandık 
 #define isleap(y)  ((y) % 4 == 0 && ((y) % 100 != 0 || (y) % 400 == 0)) 
 #define randomize() srand((unsigned)time(NULL))
 
@@ -11,9 +11,9 @@ int    isprime(int val);
 void   pline(void);
 int    ndigit(int);
 
-void set_array_random(int* p, int size);
-void print_array(const int* p, int size);
-void sort_array(int *p, int size);
+void set_array_random(int* p, size_t size);
+void print_array(const int* p, size_t size);
+void sort_array(int *p, size_t size);
 void sgets(char* p);
 void clear_input_buffer(void);
 void swap(int* px, int* py);
